CAL-587 - Memed Default para prescrições + botão de configuração

+++ timelineload.asp
@@ -308,10 +308,10 @@ SinalizarFormulariosSemPermissao = getConfig("SinalizarFormulariosSemPermissao")
                                 sqlMemed = "SELECT * FROM memedv2_prescricoes WHERE memed_id = '" & ti("MemedID") & "'"
                                 set rsMemed = db.execute(sqlMemed)
                             %>
-                                <% if cstr(session("User"))=ti("sysUser")&"" then %>
-                                    <a href="javascript:viewPrescricaoMemed(<%=ti("MemedID")%>, '<%=rsMemed("tipo")%>')">
-                                        <i class="far fa-search-plus"></i>
-                                    </a>
+                                <% if getConfig("MemedDesativada")<>1 and cstr(session("User"))=ti("sysUser")&"" then %>
+                                    <button class="btn" onclick="viewPrescricaoMemed(<%=ti("MemedID")%>, '<%=rsMemed("tipo")%>')">
+                                        Reutilizar
+                                    </button>
                                 <% end if %>
                                 <%
                                 if not rsMemed.eof then
@@ -322,7 +322,7 @@ SinalizarFormulariosSemPermissao = getConfig("SinalizarFormulariosSemPermissao")
                                     </a>
                                     <%
                                     end if
-                                    if cstr(session("User"))=ti("sysUser")&"" and aut("prescricoesX")>0  then %>
+                                    if getConfig("MemedDesativada")<>1 and cstr(session("User"))=ti("sysUser")&"" and aut("prescricoesX")>0  then %>
                                         <a href="javascript:deletePrescricaoMemed(<%=ti("id") %>, '<%=rsMemed("tipo")%>')">
                                             <i class="far fa-remove"></i>
                                         </a>
